backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
asyncpg
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
stripe
googlemaps
celery
redis
python-socketio[asyncio]
pydantic
pydantic-settings
httpx
pytest
pytest-asyncio
GeoAlchemy2
